+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.ardi.ui_ux"
+        applicationId "go.id.kominfo"
         minSdkVersion 19
         targetSdkVersion 28
         versionCode 1
