+++ app/src/main/java/go/id/kominfo/DetailProductActivity.java
@@ -1,9 +1,10 @@
-package com.ardi.ui_ux;
+package go.id.kominfo;
 
 import android.os.Build;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Window;
+
 
 import com.smarteist.autoimageslider.IndicatorAnimations;
 import com.smarteist.autoimageslider.SliderAnimations;
@@ -16,7 +17,7 @@ public class DetailProductActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_detail_product);
+        setContentView(R.layout.activity_detail_product );
         if (Build.VERSION.SDK_INT >= 21) { //ganti warna status bar diatas OS lolipop
             window = this.getWindow();
             window.setStatusBarColor(this.getResources().getColor(R.color.colorDarkPurple));
