+++ app/src/main/java/go/id/kominfo/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.ardi.ui_ux;
+package go.id.kominfo;
 
 
 import android.os.Bundle;
@@ -6,6 +6,7 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
 
 
 /**
