+++ app/src/main/java/go/id/kominfo/LoginActivity.java
@@ -1,20 +1,25 @@
-package com.ardi.ui_ux;
+package go.id.kominfo;
 
 import android.os.Build;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import android.view.Window;
 
-public class RegisterActivity extends AppCompatActivity {
+
+
+public class LoginActivity extends AppCompatActivity {
 
     Window window;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_register);
+        setContentView(R.layout.activity_login);
+
         if (Build.VERSION.SDK_INT >= 21) { //ganti warna status bar diatas OS lolipop
             window = this.getWindow();
             window.setStatusBarColor(this.getResources().getColor(R.color.colorDarkPurple));
         }
+
     }
 }
