+++ app/src/main/java/go/id/kominfo/NewProductActivity.java
@@ -1,13 +1,15 @@
-package com.ardi.ui_ux;
+package go.id.kominfo;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class SalesOrderActivity extends AppCompatActivity {
+
+
+public class NewProductActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sales_order);
+        setContentView(R.layout.activity_new_product);
     }
 }
