+++ app/src/main/java/go/id/kominfo/ProsesActivity.java
@@ -1,13 +1,14 @@
-package com.ardi.ui_ux;
+package go.id.kominfo;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class NewProductActivity extends AppCompatActivity {
+
+public class ProsesActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_new_product);
+        setContentView(R.layout.activity_proses);
     }
 }
