+++ app/src/main/res/layout/activity_detail_product.xml
@@ -8,7 +8,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".DetailProductActivity">
+    tools:context="go.id.kominfo.DetailProductActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
