+++ app/src/main/res/layout/activity_keranjang.xml
@@ -10,7 +10,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".KeranjangActivity">
+    tools:context="go.id.kominfo.KeranjangActivity">
 
     <LinearLayout
         android:id="@+id/linear_keranjang"
