+++ app/src/main/res/layout/activity_new_product.xml
@@ -9,7 +9,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".NewProductActivity">
+    tools:context="go.id.kominfo.NewProductActivity">
 
     <TextView
         android:id="@+id/tv_new_product_form"
