+++ app/src/main/res/layout/activity_proses.xml
@@ -9,7 +9,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".ProsesActivity">
+    tools:context="go.id.kominfo.ProsesActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
