+++ app/src/main/res/layout/activity_register.xml
@@ -7,7 +7,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="8dp"
-    tools:context=".LoginActivity">
+    tools:context="go.id.kominfo.LoginActivity">
 
     <RelativeLayout
         android:id="@+id/footer"
