+++ app/src/main/res/layout/activity_sales_order.xml
@@ -9,7 +9,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".SalesOrderActivity">
+    tools:context="go.id.kominfo.SalesOrderActivity">
 
     <LinearLayout
         android:layout_marginTop="@dimen/activity_horizontal_margin"
