+++ app/src/main/res/layout/fragment_home.xml
@@ -8,7 +8,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".HomeFragment">
+    tools:context="go.id.kominfo.HomeFragment">
 
     <LinearLayout
         android:layout_marginTop="24dp"
