+++ app/src/test/java/go/id/kominfo/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.ardi.ui_ux;
+package go.id.kominfo;
 
 import org.junit.Test;
